feat: 썸네일 관련 로직 추가 및 리팩토링
