+++ src/lib/clips/index.js
@@ -1,0 +1,15 @@
+const getClipsList = require("./getClipsList");
+const getUploadClipUrl = require("./getUploadClipUrl");
+const confirmUploadClip = require("./confirmUploadClip");
+const setClipTag = require("./setClipTag");
+const searchClipByTag = require("./searchClipByTag");
+const deleteClip = require("./deleteClip");
+
+module.exports = {
+    getClipsList,
+    getUploadClipUrl,
+    confirmUploadClip,
+    setClipTag,
+    searchClipByTag,
+    deleteClip,
+};
