+++ src/util/s3.js
@@ -21,35 +21,62 @@ async function getSignedUrl(fileKey, action) {
     return await S3.getSignedUrlPromise(action, params);
 }
 
+/**
+ * get signed url for upload video file
+ * @param {string} fileKey a key of the video file
+ * @returns {Promise<string>} signed url for upload video file
+ */
 async function getUploadFileUrl(fileKey) {
-    return await getSignedUrl(fileKey, "putObject");
+    return await getSignedUrl(fileKey + ".webm", "putObject");
 }
 
 /**
- * get signed url for download file
- * @param {string} fileKey a key of the file
- * @returns {Promise<string>} signed url for download file
+ * get signed url for download video file
+ * @param {string} fileKey a key of the video file
+ * @returns {Promise<string>} signed url for upload download file
  */
 async function getDownloadFileUrl(fileKey) {
-    return await getSignedUrl(fileKey, "getObject");
+    return await getSignedUrl(fileKey + ".webm", "getObject");
+}
+
+async function getUploadThumbnailUrl(fileKey) {
+    return await getSignedUrl(fileKey + ".jpg", "putObject");
 }
 
 /**
- * get signed url for upload file
- * @param {string} fileKey - a key of the file
- * @returns {Promise<string>} signed url for upload file
+ * get signed url for download thunbmail
+ * @param {string} fileKey a key of the thunbmail
+ * @returns {Promise<string>} signed url for download thunbmail
+ */
+async function getDownloadThumbnailUrl(fileKey) {
+    return await getSignedUrl(fileKey + ".jpg", "getObject");
+}
+
+/**
+ * get signed url for upload thunbmail
+ * @param {string} fileKey a key of the thunbmail
+ * @returns {Promise<string>} signed url for upload thunbmail
  */
 async function deleteFile(fileKey) {
-    const params = {
+    const params_video = {
         Bucket: S3BUCKET,
-        Key: fileKey,
+        Key: fileKey + ".webm",
+    };
+    const params_thumbnail = {
+        Bucket: S3BUCKET,
+        Key: fileKey + ".jpg",
     };
 
-    return S3.deleteObject(params).promise();
+    return {
+        video: await S3.deleteObject(params_video).promise(),
+        thumbnail: await S3.deleteObject(params_thumbnail).promise(),
+    };
 }
 
 module.exports = {
     getUploadFileUrl,
     getDownloadFileUrl,
+    getUploadThumbnailUrl,
+    getDownloadThumbnailUrl,
     deleteFile,
 };
